# Expand an optimization for short values to IgnoreCase in single-value SearchValues<string>