For values length 2 or 3, we have a special case where we don't verify matches at all if we know that false positives are impossible when all the anchors match.
https://github.com/dotnet/runtime/blob/ba9b3ba83603d1f54c00c45a87344568d4277966/src/libraries/System.Private.CoreLib/src/System/SearchValues/Strings/SingleStringSearchValuesThreeChars.cs#L362-L366
We were using this trick with case-sensitive comparisons, but it's actually also valid if we know that all the characters are ASCII letters, so this PR extends the optimization to apply to those values as well.

---

```c#
public class EarlyMatches
{
    private const string Needle = "the";
    private static readonly string s_haystack = string.Concat(Enumerable.Repeat(Needle, 10_000));
    private static readonly SearchValues<string> s_values = SearchValues.Create([Needle], StringComparison.OrdinalIgnoreCase);

    [Benchmark]
    public int Count()
    {
        int count = 0;
        ReadOnlySpan<char> haystack = s_haystack;
        while (true)
        {
            int pos = haystack.IndexOfAny(s_values);
            if (pos < 0) break;
            count++;
            haystack = haystack.Slice(Needle.Length);
        }
        return count;
    }
}
```

| Method | Toolchain | Mean     | Error    | Ratio | Code Size |
|------- |---------- |---------:|---------:|------:|----------:|
| Count  | main      | 29.41 us | 0.517 us |  1.00 |     946 B |
| Count  | pr        | 23.30 us | 0.269 us |  0.79 |     726 B |